# contracts/api/aione-core-v1.openapi.yaml
openapi: 3.1.0
info:
  title: AIONE Core Data API
  version: 1.1.0-candidate
  description: >-
    第7阶段核心数据、岗位AI办公室与AI秘书候选契约。页面不是数据库边界；工作、人才、ERP、分析和AI秘书读取同一份业务事实。
servers:
  - url: /api/v1
paths:
...
  /time-summary:
    get:
      summary: 按人员汇总每日有效工作时间
      parameters:
        - { in: query, name: personId, required: true, schema: { type: string } }
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
      responses:
        '200': { description: OK }
  /people/{personId}/work-summary:
    get:
      summary: 人才之家/工作之家共享的个人工作证据摘要
      parameters:
        - { in: path, name: personId, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
...
  /ai-secretary/execute:
    post:
      summary: AI秘书受领目标；读取/分析可以直接执行，写入动作仅产生待确认方案
      responses:
        '200': { description: OK }
        '400': { description: Invalid objective }
  /ai-secretary/confirm:
    post:
      summary: 由已认证人类负责人确认AI秘书提出的写入动作
      responses:
        '200': { description: Confirmed }
        '401': { description: Human confirmation required }